pyproject.toml
[project]
name = "GitHubStats"
version = "1.0.0"
description = "Show GitHub Stats"
authors = [{ name = "C. Bernard", email = "cfr.bernard@yahoo.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "customtkinter",
    "requests",
    "pillow"
]

[project.optional-dependencies]
dev = [
    "pyinstaller",
    "pytest",
    "mypy",
    "ruff",
    "black",
    "pre-commit",
    "types-requests"
]

[project.scripts]
GHS = "GitHubStats.__main__:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["GitHubStats"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = false
disallow_untyped_defs = false
show_error_codes = true
pretty = true
exclude = "^tests/fixtures/"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.black]
line-length = 100
target-version = ['py313']
